Further implementation of unsafe methods, move static fields to memory blocks

+++ src/main/java/dev/xdark/ssvm/execution/asm/FloatToDoubleProcessor.java
@@ -17,7 +17,7 @@ public final class FloatToDoubleProcessor implements InstructionProcessor<Abstra
 	@Override
 	public Result execute(AbstractInsnNode insn, ExecutionContext ctx) {
 		var stack = ctx.getStack();
-		stack.pushWide(new DoubleValue(stack.pop().asLong()));
+		stack.pushWide(new DoubleValue(stack.pop().asDouble()));
 		return Result.CONTINUE;
 	}
 }
